login and signup works now. Now have to add clerkId to database
